bot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "islom-sattorov",email = "slmsttrv1@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
